Login/requirements.txt
anyio @ file:///home/conda/feedstock_root/build_artifacts/anyio_1688651106312/work/dist
brotlipy==0.7.0
certifi==2020.6.20
cffi @ file:///C:/ci_311/cffi_1676423759166/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///D:/bld/click_1692312014553/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
cryptography @ file:///C:/b/abs_f4do8t8jfs/croot/cryptography_1694444424531/work
dnspython @ file:///home/conda/feedstock_root/build_artifacts/dnspython_1691598144738/work
email-validator @ file:///home/conda/feedstock_root/build_artifacts/email-validator-meta_1689752101511/work
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1692026125334/work
fastapi @ file:///home/conda/feedstock_root/build_artifacts/fastapi_1693793259089/work
h11 @ file:///home/conda/feedstock_root/build_artifacts/h11_1664132893548/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1634280454336/work
hpack==4.0.0
httpcore @ file:///home/conda/feedstock_root/build_artifacts/httpcore_1694194101801/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1619110129307/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
mkl-fft @ file:///C:/b/abs_19i1y8ykas/croot/mkl_fft_1695058226480/work
mkl-random @ file:///C:/b/abs_edwkj1_o69/croot/mkl_random_1695059866750/work
mkl-service==2.4.0
numpy==1.26.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydantic @ file:///C:/b/abs_4bum6ajkso/croot/pydantic_1686125732638/work
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1689721553971/work
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
PySocks @ file:///C:/ci_311/pysocks_1676425991111/work
python-dotenv @ file:///home/conda/feedstock_root/build_artifacts/python-dotenv-split_1693865766170/work
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
sniffio @ file:///home/conda/feedstock_root/build_artifacts/sniffio_1662051266223/work
starlette @ file:///home/conda/feedstock_root/build_artifacts/starlette-recipe_1684245096404/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1695040754690/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
uvicorn @ file:///D:/bld/uvicorn-split_1690810299287/work
win-inet-pton @ file:///C:/ci_311/win_inet_pton_1676425458225/work
